+++ CHANGELOG.md
@@ -8,8 +8,13 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
 
 ### Changed
+* [Changed --exclude-analyzer to --exclude-analyzers](https://github.com/ionide/FSharp.Analyzers.SDK/pull/196) (thanks @dawedawe!)
+* [Changed --ignore-files to --exclude-files](https://github.com/ionide/FSharp.Analyzers.SDK/pull/196) (thanks @dawedawe!)
+
+### Added
 * [Add missing TAST walkers](https://github.com/ionide/FSharp.Analyzers.SDK/pull/185) (thanks @dawedawe!)
 * [Add support for --include-analyzers to ignore all others](https://github.com/ionide/FSharp.Analyzers.SDK/pull/194) (thanks @dawedawe!)
+* [Add support for --include-files to ignore all others](https://github.com/ionide/FSharp.Analyzers.SDK/pull/196) (thanks @dawedawe!)
 
 ## [0.22.0] - 2023-12-19